% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_comparisons.R
\name{simulate_comparisons}
\alias{simulate_comparisons}
\title{Simulate comparison Vectors}
\usage{
simulate_comparisons(m, u, levels, n1, n2, overlap, previous_matches = 0)
}
\description{
Simulate comparison Vectors
}
