% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_comparisons_mm.R
\name{simulate_comparisons_mm}
\alias{simulate_comparisons_mm}
\title{Simulate comparison Vectors}
\usage{
simulate_comparisons_mm(m, u, levels, n1, n2, overlap)
}
\description{
Simulate comparison Vectors
}
