% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svabl.R
\name{svabl}
\alias{svabl}
\title{SVI}
\usage{
svabl(
  hash,
  threshold = 1e-06,
  tmax = 1000,
  fixed_iterations = NULL,
  b_init = TRUE,
  B = min(1000, hash$n2),
  k = 1,
  tau = 1,
  seed = 0,
  check_every = 10,
  store_every = check_every
)
}
\description{
SVI
}
